requirements.txt
asgiref~=3.8.1
Django~=5.1.7
Pillow~=11.1.0
sqlparse~=0.5.3
django-braces~=1.17.0
django-embed-video~=1.4.10
pymemcache==4.0.0
django-debug-toolbar~=5.1.0
redis~=5.2.1
django-redisboard==8.4.0
djangorestframework~=3.15.2
requests~=2.32.3
gunicorn==23.0.0
pip~=23.2.1
attrs~=25.3.0
certifi~=2025.1.31
typing_extensions~=4.13.1
urllib3~=2.3.0
django-js-asset~=3.1.2
idna~=3.10
charset-normalizer~=3.4.1
django-ckeditor~=6.7.2
beautifulsoup4~=4.13.3